SONAR rock vs mine Prediction
The system is a Python-based predictive model designed to classify objects as either rocks or mines using SONAR data. This model employs Logistic Regression, a statistical method for binary classification, to analyze the features of the SONAR signals and predict the type of object detected. The Logistic Regression model is trained on a labeled dataset of SONAR signals, allowing it to learn the distinguishing characteristics of rocks and mines.